requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=1.10.0
torchvision>=0.11.0

# Data processing
nibabel>=3.2.0
opencv-python>=4.5.0
scipy>=1.7.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
matplotlib>=3.5.0

# Optional but recommended
pillow>=8.3.0
scikit-image>=0.18.0
